Cleanup spec tests and ssz_static handler.
